apply more Annotation

+++ src/com/xruby/runtime/builtin/ArrayPacker.java
@@ -634,9 +634,9 @@ class ArrayPacker {
                             throw new RubyException(RubyRuntime.RuntimeErrorClass, "too few for type " + type); // #TODO: message
 
                         if (from instanceof RubyFixnum) {
-                            f = (long) (((RubyFixnum) from).intValue() & 0xffffffff);
+                            f = (long) (((RubyFixnum) from).toInt() & 0xffffffff);
                         } else if (from instanceof RubyFloat) {
-                            f = ((long) ((RubyFloat) from).doubleValue() & 0xffffffffffffffffL);
+                            f = ((long) ((RubyFloat) from).toFloat() & 0xffffffffffffffffL);
                         } else if (from instanceof RubyBignum) {
                             f = (((RubyBignum) from).getInternal().longValue() & 0xffffffffffffffffL);
                         } else {
